refactor: modifier capability를 폴더별로 분리
